Refactored tests to BaseProvider, changed travis node version to 8
