+++ src/db/provider_clients/BaseProvider.js
@@ -1,4 +1,8 @@
 // @flow
+import { promisify } from 'util';
+import { writeFile } from 'fs';
+import json2csv from 'json2csv';
+import SqliteJsonExport from 'sqlite-json-export';
 import Tunnel from '../Tunnel';
 import clients from './';
 import * as config from '../../Config';
@@ -7,6 +11,7 @@ import type { sshTunnelType } from '../Tunnel';
 import type { serverType, databaseType } from './ProviderInterface';
 
 const logger = createLogger('db');
+const writeFilePromise = promisify(writeFile);
 
 export default class BaseProvider {
 
@@ -244,5 +249,67 @@ export default class BaseProvider {
       throw new Error('There is no connection available.');
     }
     return true;
+  }
+
+  async getJsonString(exportOptions: exportOptionsType) {
+    const exporter = new SqliteJsonExport(this.connection.dbConfig.database);
+    this.checkUnsupported(exportOptions);
+
+    if (
+      'tables' in exportOptions &&
+      'table' in exportOptions
+    ) {
+      throw new Error('You cannot give both "tables" and "table". Choose one');
+    }
+
+    const getSingleTable = (tableName: string): Promise<string> =>
+      new Promise((resolve, reject) => {
+        const options = {
+          table: tableName
+        };
+        exporter.json(options, (err: Error, json: string) => {
+          if (err) return reject(err);
+          return resolve(json);
+        });
+      });
+
+    // Multiple tables
+    if ('tables' in exportOptions) {
+      const results = await Promise.all(
+          exportOptions.tables.map(tableName => getSingleTable(tableName))
+        )
+        .then(tableJsonStrings => tableJsonStrings.join(','));
+
+      return ['[', ...results, ']'].join('');
+    }
+
+    // Single table
+    return getSingleTable(exportOptions.table);
+  }
+
+  async getCsvString(exportOptions: exportOptionsType) {
+    if ('tables' in exportOptions) {
+      throw new Error('Exporting multiple tables to csv is currently not supported');
+    }
+
+    const jsonString = await this.getJsonString(exportOptions);
+    const parsedJson = JSON.parse(jsonString);
+
+    return json2csv({
+      data: parsedJson,
+      fields: Object.keys(parsedJson[0])
+    });
+  }
+
+  async exportJson(exportOptions: exportOptionsType, filename: string) {
+    const jsonString = await this.getJsonString(exportOptions);
+    await writeFilePromise(filename, jsonString);
+    return jsonString;
+  }
+
+  async exportCsv(exportOptions: exportOptionsType, filename: string) {
+    const csvString = await this.getCsvString(exportOptions);
+    await writeFilePromise(filename, csvString);
+    return csvString;
   }
 }
